// dict/api/user/types.ts
/* eslint-disable */
// tslint:disable
/**
 * MICRO SERVICES CURD OPERATIONS
 * 
 *
 * OpenAPI spec version: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface UserApi {
	userCreate: (request: Api.BODYDATA | undefined) => Promise<UserCreateResponse>
	getUser: () => Promise<GetUserResponse>
}

export type UserCreateResponse = UserCreate201Response | UserCreate400Response

export interface UserCreate201Response {
	status: 201
	body: Api.ApiResponse
	headers?: never
}

export interface UserCreate400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}

export type GetUserResponse = GetUser201Response | GetUser400Response

export interface GetUser201Response {
	status: 201
	body: Api.BODYDATA[]
	headers?: never
}

export interface GetUser400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}